meta: convert test to typescript
